% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{delete_account_setting}
\alias{delete_account_setting}
\title{Modifies the ARN and resource ID format of a resource for a specified IAM user, IAM role, or the root user for an account}
\usage{
delete_account_setting(
  name = "serviceLongArnFormat"|"taskLongArnFormat"|"containerInstanceLongArnFormat",
  principalArn = "string"
)
}
\arguments{
\item{name}{[required] The resource name for which to disable the new format. If \code{serviceLongArnFormat} is specified, the ARN for your Amazon ECS services is affected. If \code{taskLongArnFormat} is specified, the ARN and resource ID for your Amazon ECS tasks is affected. If \code{containerInstanceLongArnFormat} is specified, the ARN and resource ID for your Amazon ECS container instances is affected.}

\item{principalArn}{The ARN of the principal, which can be an IAM user, IAM role, or the root user. If you specify the root user, it modifies the ARN and resource ID format for all IAM users, IAM roles, and the root user of the account unless an IAM user or role explicitly overrides these settings for themselves. If this field is omitted, the setting are changed only for the authenticated user.}
}
\description{
Modifies the ARN and resource ID format of a resource for a specified IAM user, IAM role, or the root user for an account. You can specify whether the new ARN and resource ID format are disabled for new resources that are created.
}
\examples{

}
