% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{delete_attributes}
\alias{delete_attributes}
\title{Deletes one or more custom attributes from an Amazon ECS resource}
\usage{
delete_attributes(attributes, cluster = NULL)
}
\arguments{
\item{attributes}{The attributes to delete from your resource. You can specify up to 10 attributes per request. For custom attributes, specify the attribute name and target ID, but do not specify the value. If you specify the target ID using the short form, you must also specify the target type.}

\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that contains the resource to delete attributes. If you do not specify a cluster, the default cluster is assumed.}
}
\description{
Deletes one or more custom attributes from an Amazon ECS resource.
}
\examples{

}
