% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{delete_cluster}
\alias{delete_cluster}
\title{Deletes the specified cluster}
\usage{
delete_cluster(cluster)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster to delete.}
}
\description{
Deletes the specified cluster. You must deregister all container instances from this cluster before you may delete it. You can list the container instances in a cluster with ListContainerInstances and deregister them with DeregisterContainerInstance.
}
\examples{
# This example deletes an empty cluster in your default region.
delete_cluster(
  cluster = "my_cluster"
)

}
