% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{describe_clusters}
\alias{describe_clusters}
\title{Describes one or more of your clusters}
\usage{
describe_clusters(clusters = NULL, include = NULL)
}
\arguments{
\item{clusters}{A list of up to 100 cluster names or full cluster Amazon Resource Name (ARN) entries. If you do not specify a cluster, the default cluster is assumed.}

\item{include}{Additional information about your clusters to be separated by launch type, including:
\itemize{
\item runningEC2TasksCount
\item runningFargateTasksCount
\item pendingEC2TasksCount
\item pendingFargateTasksCount
\item activeEC2ServiceCount
\item activeFargateServiceCount
\item drainingEC2ServiceCount
\item drainingFargateServiceCount
}}
}
\description{
Describes one or more of your clusters.
}
\section{Accepted Parameters}{
\preformatted{describe_clusters(
  clusters = list(
    "string"
  ),
  include = list(
    "STATISTICS"|"TAGS"
  )
)
}
}

\examples{
# This example provides a description of the specified cluster in your
# default region.
\donttest{describe_clusters(
  clusters = list(
    "default"
  )
)}

}
