% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{describe_container_instances}
\alias{describe_container_instances}
\title{Describes Amazon Elastic Container Service container instances}
\usage{
describe_container_instances(containerInstances, cluster = NULL)
}
\arguments{
\item{containerInstances}{A list of up to 100 container instance IDs or full Amazon Resource Name (ARN) entries.}

\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that hosts the container instances to describe. If you do not specify a cluster, the default cluster is assumed.}
}
\description{
Describes Amazon Elastic Container Service container instances. Returns metadata about registered and remaining resources on each container instance requested.
}
\examples{
# This example provides a description of the specified container instance
# in your default region, using the container instance UUID as an
# identifier.
describe_container_instances(
  cluster = "default",
  containerInstances = list(
    "f2756532-8f13-4d53-87c9-aed50dc94cd7"
  )
)

}
