% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{describe_services}
\alias{describe_services}
\title{Describes the specified services running in your cluster}
\usage{
describe_services(cluster = NULL, services)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN)the cluster that hosts the service to describe. If you do not specify a cluster, the default cluster is assumed.}

\item{services}{[required] A list of services to describe. You may specify up to 10 services to describe in a single operation.}
}
\description{
Describes the specified services running in your cluster.
}
\section{Accepted Parameters}{
\preformatted{describe_services(
  cluster = "string",
  services = list(
    "string"
  )
)
}
}

\examples{
# This example provides descriptive information about the service named
# ``ecs-simple-service``.
\donttest{describe_services(
  services = list(
    "ecs-simple-service"
  )
)}

}
