% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{describe_task_definition}
\alias{describe_task_definition}
\title{Describes a task definition}
\usage{
describe_task_definition(taskDefinition)
}
\arguments{
\item{taskDefinition}{The \code{family} for the latest \code{ACTIVE} revision, \code{family} and \code{revision} (\code{family:revision}) for a specific revision in the family, or full Amazon Resource Name (ARN) of the task definition to describe.}
}
\description{
Describes a task definition. You can specify a \code{family} and \code{revision} to find information about a specific task definition, or you can simply specify the family to find the latest \code{ACTIVE} revision in that family.
}
\details{
You can only describe \code{INACTIVE} task definitions while an active task or service references them.
}
\examples{
# This example provides a description of the specified task definition.
describe_task_definition(
  taskDefinition = "hello_world:8"
)

}
