% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{describe_tasks}
\alias{describe_tasks}
\title{Describes a specified task or tasks}
\usage{
describe_tasks(cluster = NULL, tasks)
}
\arguments{
\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that hosts the task to describe. If you do not specify a cluster, the default cluster is assumed.}

\item{tasks}{[required] A list of up to 100 task IDs or full ARN entries.}
}
\description{
Describes a specified task or tasks.
}
\section{Accepted Parameters}{
\preformatted{describe_tasks(
  cluster = "string",
  tasks = list(
    "string"
  )
)
}
}

\examples{
# This example provides a description of the specified task, using the
# task UUID as an identifier.
\donttest{describe_tasks(
  tasks = list(
    "c5cba4eb-5dad-405e-96db-71ef8eefe6a8"
  )
)}

}
