% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{discover_poll_endpoint}
\alias{discover_poll_endpoint}
\title{This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent}
\usage{
discover_poll_endpoint(containerInstance = NULL, cluster = NULL)
}
\arguments{
\item{containerInstance}{The container instance ID or full ARN of the container instance. The ARN contains the \code{arn:aws:ecs} namespace, followed by the Region of the container instance, the AWS account ID of the container instance owner, the \code{container-instance} namespace, and then the container instance ID. For example, \code{arn:aws:ecs:region:aws_account_id:container-instance/container_instance_ID region:aws_account_id:container-instance/container_instance_ID }.}

\item{cluster}{The short name or full Amazon Resource Name (ARN) of the cluster that the container instance belongs to.}
}
\description{
This action is only used by the Amazon ECS agent, and it is not intended for use outside of the agent.
}
\details{
Returns an endpoint for the Amazon ECS agent to poll for updates.
}
\examples{

}
