% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{List the tags for an Amazon ECS resource}
\usage{
list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to list the tags. Currently, the supported resources are Amazon ECS tasks, services, task definitions, clusters, and container instances.}
}
\description{
List the tags for an Amazon ECS resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  resourceArn = "string"
)
}
}

\examples{
# This example lists the tags for the 'dev' cluster.
\donttest{list_tags_for_resource(
  resourceArn = "arn:aws:ecs:region:aws_account_id:cluster/dev"
)}

}
