% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_package.R
\docType{package}
\name{paws.ecs-package}
\alias{paws.ecs}
\alias{paws.ecs-package}
\title{paws.ecs: Amazon EC2 Container Service}
\description{
Amazon Elastic Container Service (Amazon ECS) is a
    highly scalable, fast, container management service that makes it easy
    to run, stop, and manage Docker containers on a cluster. You can host
    your cluster on a serverless infrastructure that is managed by Amazon
    ECS by launching your services or tasks using the Fargate launch type.
    For more control, you can host your tasks on a cluster of Amazon
    Elastic Compute Cloud (Amazon EC2) instances that you manage by using
    the EC2 launch type. For more information about launch types, see
    [Amazon ECS Launch
    Types](http://docs.aws.amazon.com/AmazonECS/latest/developerguide/launch_types.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
