% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Deletes specified tags from a resource}
\usage{
untag_resource(
  resourceArn = "string",
  tagKeys = list(
    "string"
  )
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource from which to delete tags. Currently, the supported resources are Amazon ECS tasks, services, task definitions, clusters, and container instances.}

\item{tagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a resource.
}
\examples{
# This example deletes the 'team' tag from the 'dev' cluster.
untag_resource(
  resourceArn = "arn:aws:ecs:region:aws_account_id:cluster/dev",
  tagKeys = list(
    "team"
  )
)

}
