% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ecs_operations.R
\name{describe_services}
\alias{describe_services}
\title{Describes the specified services running in your cluster}
\usage{
describe_services(services, cluster = NULL)
}
\arguments{
\item{services}{A list of services to describe. You may specify up to 10 services to describe in a single operation.}

\item{cluster}{The short name or full Amazon Resource Name (ARN)the cluster that hosts the service to describe. If you do not specify a cluster, the default cluster is assumed.}
}
\description{
Describes the specified services running in your cluster.
}
\examples{
# This example provides descriptive information about the service named
# ``ecs-simple-service``.
describe_services(
  services = list(
    "ecs-simple-service"
  )
)

}
